keeping rollout and adding comments. works.
