+++ pymdp/envs/__init__.py
@@ -1,1 +1,2 @@
 from .tmaze import TMaze
+from .rollout import rollout
