+++ pymdp/envs/aif_loop.py
@@ -2,47 +2,44 @@ import jax.numpy as jnp
 import jax.random as jr
 import jax.tree_util as jtu
 import jax.lax
-
+from .env import Env
 from pymdp.agent import Agent
-from pymdp.envs.env import Env
 
 
-def rollout(agent: Agent, env: Env, num_timesteps: int, rng_key: jr.PRNGKey, policy_search=None):
+def aif_loop(agent: Agent, env: Env, num_timesteps: int, rng_key: jr.PRNGKey):
     """
     Rollout an agent in an environment for a number of timesteps.
 
     Parameters
     ----------
-    agent: active inference agent 
-    env: environment that can step forward and return observations
-    num_timesteps: how many timesteps to simulate
-    rng_key: random key for sampling
-    policy_search: optional custom policy inference function such as sophisticated inference
-    
+    agent: ``Agent``
+        Agent to interact with the environment
+    env: ``Env`
+        Environment to interact with
+    num_timesteps: ``int``
+        Number of timesteps to rollout for
+    rng_key: ``PRNGKey``
+        Random key to use for sampling actions
+
     Returns
     ----------
     last: ``dict``
-        dictionary from the last timestep about the rollout, i.e., the final action, observation, beliefs, etc.
+        Carry dictionary from the last timestep
     info: ``dict``
-        dictionary containing information about the rollout, i.e. executed actions, observations, beliefs, etc.
+        Dictionary containing information about the rollout, i.e. executed actions, observations, beliefs, etc.
     env: ``Env``
-        environment state after the rollout
+        Environment state after the rollout
     """
-
-    # get the batch_size of the agent
+    # get the batch_size of the agent (number of parallel environments)
     batch_size = agent.batch_size
 
-    # default policy search just uses standard active inference policy selection
-    if policy_search is None:
+    def default_policy_search(agent, qs, rng_key):
+        qpi, _ = agent.infer_policies(qs)
+        return qpi, None
 
-        def default_policy_search(agent, qs, rng_key):
-            qpi, _ = agent.infer_policies(qs) # infer_policies computes posterior over policies using EFE
-            return qpi, None
-
-        policy_search = default_policy_search
+    policy_search = default_policy_search # default policy inference that returns action probabilities and no additional info
 
     def step_fn(carry, x):
-        # carrying the current timestep's action, observation, beliefs, empirical prior, environment state, and random key
         action_t = carry["action_t"]
         observation_t = carry["observation_t"]
         qs = carry["qs"]
@@ -50,30 +47,29 @@ def rollout(agent: Agent, env: Env, num_timesteps: int, rng_key: jr.PRNGKey, pol
         env = carry["env"]
         rng_key = carry["rng_key"]
 
-        # perform state inference using variational inference (FPI) - uses A matrix to map between hidden states and observations
+        # we infer the posterior using FPI so we don't need past actions or qs_hist
         qs = agent.infer_states(
             observations=observation_t,
             empirical_prior=empirical_prior,
         )
 
         rng_key, key = jr.split(rng_key)
-        qpi, _ = policy_search(agent, qs, key) # compute policy posterior using EFE - uses C to consider preferred outcomes
+        qpi, _ = policy_search(agent, qs, key)
 
         keys = jr.split(rng_key, batch_size + 1)
         rng_key = keys[0]
-        action_t = agent.sample_action(qpi, rng_key=keys[1:]) # sample action from policy distribution
+        action_t = agent.sample_action(qpi, rng_key=keys[1:])
 
         keys = jr.split(rng_key, batch_size + 1)
         rng_key = keys[0]
-        observation_t, env = env.step(rng_key=keys[1:], actions=action_t) # step environment forward with chosen action
+        observation_t, env = env.step(rng_key=keys[1:], actions=action_t)
 
-        empirical_prior, qs = agent.update_empirical_prior(action_t, qs) # updating the prior over hidden states and using B matrix to predict next state given action
+        empirical_prior, qs = agent.update_empirical_prior(action_t, qs)
 
-        # carrying the next timestep's action, observation, beliefs, empirical prior, environment state, and random key
         carry = {
             "action_t": action_t,
             "observation_t": observation_t,
-            "qs": jtu.tree_map(lambda x: x[:, -1:, ...], qs), # keep only latest belief
+            "qs": jtu.tree_map(lambda x: x[:, -1:, ...], qs),
             "empirical_prior": empirical_prior,
             "env": env,
             "rng_key": rng_key,
@@ -88,22 +84,21 @@ def rollout(agent: Agent, env: Env, num_timesteps: int, rng_key: jr.PRNGKey, pol
 
         return carry, info
 
-    # initialise first observation from environment
+    # generate initial observation
     keys = jr.split(rng_key, batch_size + 1)
     rng_key = keys[0]
     observation_0, env = env.reset(keys[1:])
 
-    # specify initial beliefs using D 
+    # initial belief
     qs_0 = jtu.tree_map(lambda x: jnp.expand_dims(x, -2), agent.D)
 
-    # get initial policy and action distribution - unused, just for shape matching
+    # infer initial action to get the right shape
     qpi_0, _ = agent.infer_policies(qs_0)
     keys = jr.split(rng_key, batch_size + 1)
     rng_key = keys[0]
     action_t = agent.sample_action(qpi_0, rng_key=keys[1:])
-    action_t *= 0 # zero out initial action as no action taken yet
+    action_t *= 0
 
-    # set up initial state to carry through timesteps
     initial_carry = {
         "qs": qs_0,
         "action_t": action_t,
@@ -113,10 +108,9 @@ def rollout(agent: Agent, env: Env, num_timesteps: int, rng_key: jr.PRNGKey, pol
         "rng_key": rng_key,
     }
 
-    # run the active inference loop for num_timesteps using jax.lax.scan
+    # scan over time dimension
     last, info = jax.lax.scan(step_fn, initial_carry, jnp.arange(num_timesteps))
 
-    # prepare initial info to concatenate with trajectory
     initial_info = {
         "action": jnp.expand_dims(action_t, 0),
         "observation": [jnp.expand_dims(o, 0) for o in observation_0],  
@@ -125,7 +119,6 @@ def rollout(agent: Agent, env: Env, num_timesteps: int, rng_key: jr.PRNGKey, pol
         "env": env
     }
     
-    # helper function to concatenate initial state with trajectory by dealing with different shapes and data types
     def concat_or_pass(init, steps):
         if isinstance(init, list):
             return [jnp.concatenate([i, s], axis=0) for i, s in zip(init, steps)]
@@ -137,7 +130,6 @@ def rollout(agent: Agent, env: Env, num_timesteps: int, rng_key: jr.PRNGKey, pol
             return jnp.concatenate([init, steps], axis=0)
         return steps
 
-    # combine initial info with trajectory info
     info = jtu.tree_map(concat_or_pass, initial_info, info)
 
     return last, info, env
