+++ pymdp/envs/tmaze.py
@@ -1,9 +1,6 @@
 import os
 import math
 import jax.numpy as jnp
-import jax.random as jr
-import jax.tree_util as jtu
-import jax.lax
 import matplotlib.pyplot as plt
 import matplotlib.patches as patches
 from matplotlib.offsetbox import OffsetImage, AnnotationBbox
@@ -11,8 +8,7 @@ import scipy.ndimage as ndimage
 from pymdp.utils import fig2img
 from equinox import field
 from .env import Env
-from pymdp.agent import Agent
-# import io
+
 
 # load assets
 assets_dir = os.path.join(os.path.dirname(os.path.realpath(__file__)), "assets")
@@ -204,8 +200,16 @@ class TMaze(Env):
         D.append(D_reward)
         return D
 
-    def render(self, mode="human"):
-        batch_size = self.params["A"][0].shape[0]
+    def render(self, mode="human", observations=None):
+        if observations is not None:
+            current_obs = observations
+            batch_size = observations[0].shape[0]
+        else:
+            current_obs = self.current_obs
+            batch_size = self.params["A"][0].shape[0]
+
+        plt.clf()  # Clear the current figure
+        
 
         # create n x n subplots for the batch_size
         n = math.ceil(math.sqrt(batch_size))
@@ -272,7 +276,7 @@ class TMaze(Env):
             )
 
             # show the cue
-            cue = self.current_obs[2][i, 0]
+            cue = current_obs[2][i, 0]
             if cue == 0:
                 cue_color = "tab:gray"
                 edge_color = "tab:gray"
@@ -296,10 +300,8 @@ class TMaze(Env):
             )
 
             # show the reward
-            loc = self.current_obs[0][i, 0]
-
-            reward = self.current_obs[1][i, 0]
-
+            loc = current_obs[0][i, 0]
+            reward = current_obs[1][i, 0]
             if loc == 1:
                 coords = (0.5, 0.5)
             elif loc == 2:
@@ -351,7 +353,7 @@ class TMaze(Env):
                 ab_mouse = ax.add_artist(ab_mouse)
                 ab_mouse.set_zorder(3)
 
-        # Hide any extra subplots if batch_size isn't a perfect square
+        # hide any extra subplots if batch_size isn't a perfect square
         for i in range(batch_size, n * n):
             fig.delaxes(axes.flatten()[i])
 
@@ -360,143 +362,6 @@ class TMaze(Env):
         if mode == "human":
             plt.show()
         elif mode == "rgb_array":
-            return fig2img(fig)
-
-
-
-def aif_loop(agent: Agent, env: Env, num_timesteps: int, rng_key: jr.PRNGKey):
-    """
-    Rollout an agent in an environment for a number of timesteps.
-
-    Parameters
-    ----------
-    agent: ``Agent``
-        Agent to interact with the environment
-    env: ``Env`
-        Environment to interact with
-    num_timesteps: ``int``
-        Number of timesteps to rollout for
-    rng_key: ``PRNGKey``
-        Random key to use for sampling actions
-
-    Returns
-    ----------
-    last: ``dict``
-        Carry dictionary from the last timestep
-    info: ``dict``
-        Dictionary containing information about the rollout, i.e. executed actions, observations, beliefs, etc.
-    env: ``Env``
-        Environment state after the rollout
-    """
-    # get the batch_size of the agent
-    batch_size = agent.batch_size
-
-    def default_policy_search(agent, qs, rng_key):
-        qpi, _ = agent.infer_policies(qs)
-        return qpi, None
-
-    policy_search = default_policy_search
-
-    def step_fn(carry, x):
-        action_t = carry["action_t"]
-        observation_t = carry["observation_t"]
-        qs = carry["qs"]
-        empirical_prior = carry["empirical_prior"]
-        env = carry["env"]
-        rng_key = carry["rng_key"]
-
-        # we infer the posterior using FPI so we don't need past actions or qs_hist
-        qs = agent.infer_states(
-            observations=observation_t,
-            empirical_prior=empirical_prior,
-        )
-
-        rng_key, key = jr.split(rng_key)
-        qpi, _ = policy_search(agent, qs, key)
-
-        keys = jr.split(rng_key, batch_size + 1)
-        rng_key = keys[0]
-        action_t = agent.sample_action(qpi, rng_key=keys[1:])
-
-        keys = jr.split(rng_key, batch_size + 1)
-        rng_key = keys[0]
-        observation_t, env = env.step(rng_key=keys[1:], actions=action_t)
-
-        empirical_prior, qs = agent.update_empirical_prior(action_t, qs)
-
-        carry = {
-            "action_t": action_t,
-            "observation_t": observation_t,
-            "qs": jtu.tree_map(lambda x: x[:, -1:, ...], qs),
-            "empirical_prior": empirical_prior,
-            "env": env,
-            "rng_key": rng_key,
-        }
-        info = {
-            "qpi": qpi,
-            "qs": jtu.tree_map(lambda x: x[:, 0, ...], qs),
-            "env": env,
-            "observation": observation_t,
-            "action": action_t,
-        }
-
-        return carry, info
-
-    # generate initial observation
-    keys = jr.split(rng_key, batch_size + 1)
-    rng_key = keys[0]
-    observation_0, env = env.reset(keys[1:])
-
-    # initial belief
-    qs_0 = jtu.tree_map(lambda x: jnp.expand_dims(x, -2), agent.D)
-
-    # infer initial action to get the right shape
-    qpi_0, _ = agent.infer_policies(qs_0)
-    keys = jr.split(rng_key, batch_size + 1)
-    rng_key = keys[0]
-    action_t = agent.sample_action(qpi_0, rng_key=keys[1:])
-    action_t *= 0
-
-    initial_carry = {
-        "qs": qs_0,
-        "action_t": action_t,
-        "observation_t": observation_0,
-        "empirical_prior": agent.D,
-        "env": env,
-        "rng_key": rng_key,
-    }
-
-    # scan over time dimension
-    last, info = jax.lax.scan(step_fn, initial_carry, jnp.arange(num_timesteps))
-
-    initial_info = {
-        "action": jnp.expand_dims(action_t, 0),
-        "observation": [jnp.expand_dims(o, 0) for o in observation_0],  
-        "qs": jtu.tree_map(lambda x: jnp.transpose(x, (1, 0) + tuple(range(2, x.ndim))), qs_0), 
-        "qpi": jnp.expand_dims(qpi_0, 0),  
-        "env": env
-    }
-    # def concat_or_pass(init, steps):
-    #     if isinstance(init, list):
-    #         return [jnp.concatenate([i, s], axis=0) for i, s in zip(init, steps)]
-    #     elif isinstance(init, jnp.ndarray):
-    #         if init.ndim < steps.ndim:
-    #             init = jnp.expand_dims(init, 1)
-    #         return jnp.concatenate([init, steps], axis=0)
-    #     return steps
-    
-    def concat_or_pass(init, steps):
-        if isinstance(init, list):
-            return [jnp.concatenate([i, s], axis=0) for i, s in zip(init, steps)]
-        elif isinstance(init, jnp.ndarray):
-            if init.ndim < steps.ndim:
-                init = jnp.expand_dims(init, 0)  # Add time dimension at start
-            elif init.shape[1:] != steps.shape[1:]:  # If dimensions after time don't match
-                # Transpose to match steps dimensions
-                init = jnp.transpose(init, (1, 0) + tuple(range(2, init.ndim)))
-            return jnp.concatenate([init, steps], axis=0)
-        return steps
-
-    info = jtu.tree_map(concat_or_pass, initial_info, info)
-
-    return last, info, env
+            img = fig2img(fig)
+            plt.close(fig) 
+            return img
